requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-common-sql==1.7.1
apache-airflow-providers-http==4.5.1
kafka-python==2.0.2
clickhouse-driver==0.2.6
requests==2.31.0
python-dateutil==2.8.2
PyYAML==6.0.1
SQLAlchemy==1.4.49
psutil==5.9.5
gunicorn==21.2.0